Extract song component
