+++ app/components/Song/Song.module.scss
@@ -1,0 +1,3 @@
+.song {
+  
+}
